requirements.txt
psycopg2-binary==2.9.3
typer==0.5.0
